rich link component
